```
goos: linux
goarch: amd64
pkg: github.com/redis/rueidis/rueidislimiter
cpu: AMD Ryzen 7 7840HS with Radeon 780M Graphics   
          │   old.txt   │              new.txt               │
          │   sec/op    │   sec/op     vs base               │
AllowN-16   872.9n ± 1%   787.0n ± 0%  -9.84% (p=0.000 n=20)

          │  old.txt   │              new.txt               │
          │    B/op    │    B/op     vs base                │
AllowN-16   440.0 ± 0%   368.0 ± 0%  -16.36% (p=0.000 n=20)

          │   old.txt   │              new.txt               │
          │  allocs/op  │ allocs/op   vs base                │
AllowN-16   12.000 ± 0%   8.000 ± 0%  -33.33% (p=0.000 n=20)
```

```
go test -coverprofile=coverage.out && go tool cover -func=coverage.out
PASS
coverage: 100.0% of statements
ok      github.com/redis/rueidis/rueidislimiter 0.004s
github.com/redis/rueidis/rueidislimiter/limit.go:10:    WithCustomRateLimit     100.0%
github.com/redis/rueidis/rueidislimiter/limiter.go:53:  NewRateLimiter          100.0%
github.com/redis/rueidis/rueidislimiter/limiter.go:84:  Limit                   100.0%
github.com/redis/rueidis/rueidislimiter/limiter.go:88:  Check                   100.0%
github.com/redis/rueidis/rueidislimiter/limiter.go:92:  Allow                   100.0%
github.com/redis/rueidis/rueidislimiter/limiter.go:96:  AllowN                  100.0%
github.com/redis/rueidis/rueidislimiter/syncp.go:15:    Capacity                100.0%
github.com/redis/rueidis/rueidislimiter/syncp.go:19:    ResetLen                100.0%
total:                                                  (statements)            100.0%
```

Closes #775